pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "persona-steering"
version = "0.1.0"
description = "Steer LLM behavior with extracted persona vectors"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["llm", "persona", "steering", "transformers", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "accelerate",
    "safetensors",
    "pyyaml",
    "click",
    "rich",
    "prompt_toolkit",
    "jsonlines",
    "numpy",
]

[project.optional-dependencies]
structured = [
    "outlines>=0.1.0",  # Structured JSON generation
]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]

[project.scripts]
persona-steer = "src.cli:main"

[project.urls]
Homepage = "https://github.com/user/persona-steering"
Documentation = "https://github.com/user/persona-steering#readme"
Repository = "https://github.com/user/persona-steering"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
